requirements.txt
asgi-redis==1.0.0
asgiref==2.2.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
channels==2.0.2
constantly==15.1.0
daphne==2.1.0
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.7.7
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack-python==0.5.6
pytz==2018.3
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.9.0
zope.interface==4.4.3
